requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
solana>=0.25.0
sqlalchemy>=1.4.0
slowapi>=0.1.0
pydantic>=1.10.0
httpx>=0.23.0
solders>=0.25.0
python-jose[cryptography]>=3.3.0
databases[postgresql]>=0.5.0
asyncpg>=0.25.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
streamlit>=1.22.0